+++ vllm/renderers/inputs/__init__.py
@@ -1,0 +1,19 @@
+# SPDX-License-Identifier: Apache-2.0
+# SPDX-FileCopyrightText: Copyright contributors to the vLLM project
+from .protocol import (
+    DecoderDictPrompt,
+    DecoderOnlyDictPrompt,
+    DictPromptType,
+    EncoderDecoderDictPrompt,
+    EncoderDictPrompt,
+    SingletonDictPrompt,
+)
+
+__all__ = [
+    "DecoderOnlyDictPrompt",
+    "EncoderDictPrompt",
+    "DecoderDictPrompt",
+    "EncoderDecoderDictPrompt",
+    "SingletonDictPrompt",
+    "DictPromptType",
+]
