+++ vllm/v1/engine/utils.py
@@ -18,10 +18,10 @@ import zmq
 from vllm import envs
 from vllm.config import CacheConfig, ParallelConfig, VllmConfig
 from vllm.inputs import PromptType
-from vllm.inputs.parse import get_prompt_components
 from vllm.logger import init_logger
 from vllm.platforms import current_platform
 from vllm.ray.ray_env import get_env_vars_to_copy
+from vllm.renderers.inputs.parse import get_prompt_components
 from vllm.utils.network_utils import get_open_zmq_ipc_path, zmq_socket_ctx
 from vllm.utils.system_utils import get_mp_context
 from vllm.v1.engine.coordinator import DPCoordinator
